moved component to components folder
