+++ frontend/src/main.jsx
@@ -1,14 +1,12 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import UserPage from "./UserProfile.jsx";
+import UserPage from "./components/UserProfile.jsx";
 import HomePage from "./components/HomePage";
 import "./style.css";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
     <UserPage />
-  </React.StrictMode>
-);
-    <HomePage />
+    {/* <HomePage /> */}
   </React.StrictMode>
 );
